fixed core module and move services to shared module
