+++ messenger/src/app/modules/chat/components/chat-room/chat-room.component.ts
@@ -9,11 +9,11 @@ import {
   ViewChild,
   ViewChildren
 } from '@angular/core';
-import {AuthService} from "../../../../core/services/auth.service";
-import {Chat, ChatMessage, ChatService} from "../../../../core/services/chat.service";
+import {AuthService} from "../../../../shared/services/auth.service";
+import {Chat, ChatMessage, ChatService} from "../../../../shared/services/chat.service";
 import {ActivatedRoute, Params, Router} from "@angular/router";
 import {Subject, take, takeUntil} from "rxjs";
-import {User, UsersService} from "../../../../core/services/users.service";
+import {User, UsersService} from "../../../../shared/services/users.service";
 
 @Component({
   selector: 'app-chat-room',
